Add ability to start/end DND
Add functions to start and end snooze/DND. The user can choose betwen 15, 30, 60, 90, 120, 240 and 480 minutes of "snooze time". If the user clicks on the "Do not disturb" button, snooze/DND start is triggered with a 90 minute timeout. Snooze/DND is ended when the user switches to the "Home Office" status.